Stanley's industrial attachment application
This pull request is my application for the Swahilipot industrial attachment position.

Find more about me, my favorite programming languages and the tech stack that best describes me.

Connect with me through my Twitter or LinkedIn and take a look at my current CV and portfolio.

It's amazing to contribute and engage with the Swahilipot community, and I can't wait particularly to be a part of the Fursa project which aligns with my long-term mission of ensuring more youth get tech skills that guarantee job placements or freelance work.

I look forward to being a part of the community.

Many Thanks,
Stanley Thuita.